Refactor breakout and CCXT helper types into separate modules
