+++ data/exchanges/ccxt_futures_client.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import logging
 from datetime import datetime, timezone
-from typing import Any, Callable, Protocol, TypedDict, cast, runtime_checkable
+from typing import Any, Callable, cast
 
 import pandas as pd
 
@@ -20,6 +20,7 @@ from constants import (
 from domain.abstract.exchange_client import ExchangeClient
 from domain.enums.exchange import Exchange
 from domain.enums.timeframe import Timeframe
+from data.exchanges.ccxt_types import CcxtClientOptions, CcxtFuturesApi, CcxtOpenInterestApi
 from utils.retry import RetryExhaustedError, run_with_retry
 
 try:
@@ -27,32 +28,6 @@ try:
 except ImportError:  # pragma: no cover
     ccxt = None
 
-
-class CcxtClientOptions(TypedDict):
-    defaultType: str
-
-
-class CcxtFuturesApi(Protocol):
-    markets: dict[str, dict[str, object]]
-
-    def load_markets(self) -> object:
-        ...
-
-    def fetch_ohlcv(self, symbol: str, timeframe: str, since: int, limit: int) -> list[list[float]]:
-        ...
-
-
-@runtime_checkable
-class CcxtOpenInterestApi(Protocol):
-    def fetch_open_interest_history(
-        self,
-        symbol: str,
-        timeframe: str,
-        since: int,
-        limit: int,
-        params: dict[str, str] | None = None,
-    ) -> list[dict[str, object]]:
-        ...
 
 # region Private
