+++ strategy/breakout/pending_breakout.py
@@ -1,0 +1,19 @@
+"""Pending breakout state model for breakout strategy."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+import pandas as pd
+
+from domain.enums.position_side import PositionSide
+from domain.models.level import Level
+
+
+@dataclass(slots=True)
+class PendingBreakout:
+    breakout_idx: int
+    level: Level
+    breakout_extreme: float
+    side: PositionSide
+    level_start_time: pd.Timestamp
